Agent Default Billing Department Feature
Files associated with allowing a Default Billing Department to be set
for an agent. Includes a new schema file in /InstallBundle/data/ that
reflects changes to the People Table (Person Entity) which has a Foreign
Key relationship to the Department Entity. No other "build" work was
done so may give an error about mismatch file<->database version.